archive/privacy-filter/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "av>=14.4.0",
    "face-recognition>=1.3.0",
    "fastapi[standard]>=0.116.1",
    "faster-whisper>=1.1.1",
    "llama-cpp-python>=0.3.14",
    "opencv-contrib-python-headless>=4.11.0.86",
    "silero-vad>=5.1.2",
    "soundfile>=0.13.1",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "watchfiles>=1.1.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.1",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]
